Varie in QueryEngine, SQLGenerators ecc.

+++ Source/iORM.DBBuilder.SqlGenerator.SqLite.pas
@@ -252,17 +252,17 @@ begin
     ioMdFloat:
       Result := 'REAL';
     ioMdDate:
-      Result := 'NUMERIC';
+      Result := 'REAL'; // But in SQLite documentation has NUMERIC affinity (Mauri: 02/08/2021)
     ioMdTime:
-      Result := 'NUMERIC';
+      Result := 'REAL'; // But in SQLite documentation has NUMERIC affinity (Mauri: 02/08/2021)
     ioMdDateTime:
-      Result := 'NUMERIC';
+      Result := 'REAL'; // But in SQLite documentation has NUMERIC affinity (Mauri: 02/08/2021)
     ioMdDecimal:
       Result := 'NUMERIC';
     ioMdNumeric:
       Result := 'NUMERIC';
     ioMdBoolean:
-      Result := 'NUMERIC';
+      Result := 'INTEGER'; // But in SQLite documentation has NUMERIC affinity (Mauri: 02/08/2021)
     ioMdBinary:
       Result := 'BLOB';
     ioMdCustomFieldType:
